Move all styles into assets folder
