+++ src/assets/styles/molecules/NavStyling.js
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 import { NavLink } from "react-router-dom";
-import { Solid, media, type, Gradient } from "../../index.js";
+import { Solid, media, type, Gradient } from "../../../components/index.js";
 
 export const StyledNav = styled.div`
 display: none;
