+++ src/assets/styles/templates/AddTeammatesStyling.js
