+++ src/assets/styles/templates/AppParticipantTeams.js
