+++ src/assets/styles/templates/CreateTeamStyling.js
