+++ src/assets/styles/templates/HackathonProjectsStyling.js
