+++ src/assets/styles/templates/HackthonSingleProjectStyling.js
