+++ src/assets/styles/templates/TeamViewStyling.js
